Modularize web application components
