+++ stratz_scraper/web/app.py
@@ -1,479 +1,83 @@
+"""Flask application factory and route definitions."""
+
 from __future__ import annotations
 
-from concurrent.futures import ThreadPoolExecutor
-from datetime import datetime, timedelta, timezone
-from pathlib import Path
-import traceback
+from typing import Iterable, List
 
 from flask import Flask, Response, abort, jsonify, render_template, request
 
-from ..database import (
-    db_connection,
-    retryable_execute,
-    retryable_executemany,
-    release_incomplete_assignments,
-)
-from ..heroes import HEROES, HERO_SLUGS, hero_slug
+from ..database import db_connection, retryable_execute, release_incomplete_assignments
+from ..heroes import HEROES
+from .assignment import assign_next_task
+from .config import STATIC_DIR, TEMPLATE_DIR
+from .leaderboard import fetch_best_payload, fetch_hero_leaderboard
+from .progress import fetch_progress
+from .request_utils import is_local_request
+from .seed import seed_players
+from .submissions import submit_discover_submission, submit_hero_submission
+from .tasks import reset_player_task
 
-BASE_DIR = Path(__file__).resolve().parent
-STATIC_DIR = BASE_DIR / "static"
-TEMPLATE_DIR = BASE_DIR / "templates"
-
-ASSIGNMENT_CLEANUP_KEY = "last_assignment_cleanup"
-ASSIGNMENT_CLEANUP_INTERVAL = timedelta(seconds=60)
-
-BACKGROUND_EXECUTOR = ThreadPoolExecutor(max_workers=4)
+__all__ = ["create_app"]
 
 
-def _parse_sqlite_timestamp(value: str | None) -> datetime | None:
-    if not value:
-        return None
-    try:
-        parsed = datetime.fromisoformat(value)
-    except (TypeError, ValueError):
-        return None
-    if parsed.tzinfo is None:
-        parsed = parsed.replace(tzinfo=timezone.utc)
-    return parsed
-
-
-def record_task_duration(
-    cur, steam_account_id: int, task_type: str, assigned_at_value: str | None
-) -> float | None:
-    submitted_at = datetime.now(timezone.utc)
-    assigned_at = _parse_sqlite_timestamp(assigned_at_value)
-    duration_seconds = None
-    if assigned_at is not None:
-        duration_seconds = (submitted_at - assigned_at).total_seconds()
-    retryable_execute(
-        cur,
-        """
-        INSERT INTO task_durations (
-            steamAccountId,
-            task_type,
-            assigned_at,
-            submitted_at,
-            duration_seconds
-        ) VALUES (?, ?, ?, ?, ?)
-        """,
-        (
-            steam_account_id,
-            task_type,
-            assigned_at.isoformat() if assigned_at is not None else assigned_at_value,
-            submitted_at.isoformat(),
-            duration_seconds,
-        ),
-    )
-    if duration_seconds is not None:
-        print(
-            f"[task-duration] {task_type} for {steam_account_id} took {duration_seconds:.2f}s",
-            flush=True,
-        )
-    else:
-        print(
-            f"[task-duration] {task_type} for {steam_account_id} has no assignment timestamp",
-            flush=True,
-        )
-    return duration_seconds
-
-
-def _unmark_hero_task(steam_account_id: int) -> None:
-    try:
-        with db_connection(write=True) as conn:
-            cur = conn.cursor()
-            retryable_execute(
-                cur,
-                """
-                UPDATE players
-                SET hero_done=0,
-                    hero_refreshed_at=NULL,
-                    assigned_to=NULL,
-                    assigned_at=NULL
-                WHERE steamAccountId=?
-                """,
-                (steam_account_id,),
-            )
-    except Exception:
-        traceback.print_exc()
-
-
-def _unmark_discover_task(steam_account_id: int) -> None:
-    try:
-        with db_connection(write=True) as conn:
-            cur = conn.cursor()
-            retryable_execute(
-                cur,
-                """
-                UPDATE players
-                SET discover_done=0,
-                    assigned_to=NULL,
-                    assigned_at=NULL
-                WHERE steamAccountId=?
-                """,
-                (steam_account_id,),
-            )
-    except Exception:
-        traceback.print_exc()
-
-
-def _process_hero_submission(
-    steam_account_id: int,
-    hero_stats_rows: list[tuple[int, int, int, int]],
-    best_rows: list[tuple[int, str, int, int, int]],
-    assigned_at_value: str | None,
-) -> None:
-    try:
-        with db_connection(write=True) as conn:
-            cur = conn.cursor()
-            retryable_executemany(
-                cur,
-                """
-                INSERT INTO hero_stats (steamAccountId, heroId, matches, wins)
-                VALUES (?,?,?,?)
-                ON CONFLICT(steamAccountId, heroId) DO UPDATE SET
-                    matches = CASE
-                        WHEN excluded.matches > hero_stats.matches
-                        THEN excluded.matches
-                        ELSE hero_stats.matches
-                    END,
-                    wins = CASE
-                        WHEN excluded.matches > hero_stats.matches
-                        THEN excluded.wins
-                        ELSE hero_stats.wins
-                    END
-                """,
-                hero_stats_rows or [],
-            )
-            if best_rows:
-                retryable_executemany(
-                    cur,
-                    """
-                    INSERT INTO best (hero_id, hero_name, player_id, matches, wins)
-                    VALUES (?,?,?,?,?)
-                    ON CONFLICT(hero_id) DO UPDATE SET
-                        matches=excluded.matches,
-                        wins=excluded.wins,
-                        player_id=excluded.player_id
-                    WHERE excluded.matches > best.matches
-                    """,
-                    best_rows,
-                )
-            retryable_execute(
-                cur,
-                """
-                UPDATE players
-                SET hero_done=1,
-                    hero_refreshed_at=CURRENT_TIMESTAMP
-                WHERE steamAccountId=?
-                """,
-                (steam_account_id,),
-            )
-            record_task_duration(
-                cur,
-                steam_account_id,
-                "fetch_hero_stats",
-                assigned_at_value,
-            )
-    except Exception:
-        print(
-            f"[submit-background] failed to process hero stats for {steam_account_id}",
-            flush=True,
-        )
-        traceback.print_exc()
-        _unmark_hero_task(steam_account_id)
-
-
-def _process_discover_submission(
-    steam_account_id: int,
-    discovered_ids: list[int],
-    next_depth_value: int,
-    assigned_at_value: str | None,
-) -> None:
-    try:
-        with db_connection(write=True) as conn:
-            cur = conn.cursor()
-            child_rows = [
-                (new_id, next_depth_value)
-                for new_id in discovered_ids
-                if new_id != steam_account_id
-            ]
-            if child_rows:
-                retryable_executemany(
-                    cur,
-                    """
-                    INSERT INTO players (
-                        steamAccountId,
-                        depth,
-                        hero_done,
-                        discover_done
-                    )
-                    VALUES (?,?,0,0)
-                    ON CONFLICT(steamAccountId) DO UPDATE SET
-                        depth=excluded.depth
-                    """,
-                    child_rows,
-                )
-            retryable_execute(
-                cur,
-                """
-                UPDATE players
-                SET discover_done=1,
-                    assigned_to=NULL,
-                    assigned_at=NULL
-                WHERE steamAccountId=?
-                """,
-                (steam_account_id,),
-            )
-            record_task_duration(
-                cur,
-                steam_account_id,
-                "discover_matches",
-                assigned_at_value,
-            )
-    except Exception:
-        print(
-            f"[submit-background] failed to process discovery for {steam_account_id}",
-            flush=True,
-        )
-        traceback.print_exc()
-        _unmark_discover_task(steam_account_id)
-
-
-def maybe_run_assignment_cleanup(conn) -> bool:
-    cur = conn.cursor()
-    now = datetime.now(timezone.utc)
-    last_cleanup_row = cur.execute(
-        "SELECT value FROM meta WHERE key=?",
-        (ASSIGNMENT_CLEANUP_KEY,),
-    ).fetchone()
-    if last_cleanup_row:
+def _extract_hero_rows(steam_account_id: int, heroes_payload: Iterable[dict]) -> tuple[List[tuple[int, int, int, int]], List[tuple[int, str, int, int, int]]]:
+    hero_stats_rows: List[tuple[int, int, int, int]] = []
+    best_rows: List[tuple[int, str, int, int, int]] = []
+    for hero in heroes_payload:
         try:
-            last_cleanup = datetime.fromisoformat(last_cleanup_row["value"])
+            hero_id = int(hero["heroId"])
+            matches_value = hero.get("matches", hero.get("games"))
+            if matches_value is None:
+                continue
+            matches = int(matches_value)
+            wins = int(hero.get("wins", 0))
+        except (KeyError, TypeError, ValueError):
+            continue
+        hero_stats_rows.append((steam_account_id, hero_id, matches, wins))
+        hero_name = HEROES.get(hero_id)
+        if hero_name:
+            best_rows.append((hero_id, hero_name, steam_account_id, matches, wins))
+    return hero_stats_rows, best_rows
+
+
+def _extract_discovered_ids(values: Iterable[int]) -> List[int]:
+    discovered_ids: List[int] = []
+    seen_ids: set[int] = set()
+    for value in values:
+        try:
+            candidate_id = int(value)
+        except (TypeError, ValueError):
+            continue
+        if candidate_id in seen_ids:
+            continue
+        seen_ids.add(candidate_id)
+        discovered_ids.append(candidate_id)
+    return discovered_ids
+
+
+def _resolve_next_depth(data: dict, assignment_row) -> int:
+    provided_next_depth = data.get("nextDepth")
+    if provided_next_depth is not None:
+        try:
+            return int(provided_next_depth)
         except (TypeError, ValueError):
             pass
-        else:
-            if last_cleanup.tzinfo is None:
-                last_cleanup = last_cleanup.replace(tzinfo=timezone.utc)
-            if now - last_cleanup < ASSIGNMENT_CLEANUP_INTERVAL:
-                return False
-    release_incomplete_assignments(existing=conn)
-    retryable_execute(
-        cur,
-        """
-        INSERT INTO meta (key, value)
-        VALUES (?, ?)
-        ON CONFLICT(key) DO UPDATE SET value=excluded.value
-        """,
-        (ASSIGNMENT_CLEANUP_KEY, now.isoformat()),
-    )
-    return True
-
-
-def assign_next_task(*, run_cleanup: bool = True) -> dict | None:
-    task_payload: dict | None = None
-    should_checkpoint = False
-
-    with db_connection(write=True) as conn:
-        if run_cleanup:
-            maybe_run_assignment_cleanup(conn)
-        cur = conn.cursor()
-
-        def assign_discovery() -> dict | None:
-            assigned = retryable_execute(
-                cur,
-                """
-                WITH candidate AS (
-                    SELECT steamAccountId, depth
-                    FROM players
-                    WHERE hero_done=1
-                      AND discover_done=0
-                      AND assigned_to IS NULL
-                    ORDER BY COALESCE(depth, 0) ASC, steamAccountId ASC
-                    LIMIT 1
-                )
-                UPDATE players
-                SET assigned_to='discover',
-                    assigned_at=CURRENT_TIMESTAMP
-                WHERE steamAccountId IN (SELECT steamAccountId FROM candidate)
-                  AND assigned_to IS NULL
-                RETURNING steamAccountId, depth
-                """,
-            ).fetchone()
-            if not assigned:
-                assigned = retryable_execute(
-                    cur,
-                    """
-                    WITH candidate AS (
-                        SELECT steamAccountId, depth
-                        FROM players
-                        WHERE hero_done=1
-                          AND discover_done=0
-                          AND assigned_to='discover'
-                        ORDER BY COALESCE(depth, 0) ASC, steamAccountId ASC
-                        LIMIT 1
-                    )
-                    UPDATE players
-                    SET assigned_to='discover',
-                        assigned_at=CURRENT_TIMESTAMP
-                    WHERE steamAccountId IN (SELECT steamAccountId FROM candidate)
-                      AND assigned_to='discover'
-                    RETURNING steamAccountId, depth
-                    """,
-                ).fetchone()
-            if not assigned:
-                return None
-            depth_value = assigned["depth"]
-            return {
-                "type": "discover_matches",
-                "steamAccountId": int(assigned["steamAccountId"]),
-                "depth": int(depth_value) if depth_value is not None else 0,
-            }
-
-        def restart_discovery_cycle() -> bool:
-            retryable_execute(
-                cur,
-                """
-                UPDATE players
-                SET discover_done=0,
-                    depth=CASE WHEN depth=0 THEN 0 ELSE NULL END,
-                    assigned_at=CASE WHEN assigned_to='discover' THEN NULL ELSE assigned_at END,
-                    assigned_to=CASE WHEN assigned_to='discover' THEN NULL ELSE assigned_to END
-                """,
-            )
-            return True
-
-        counter_row = cur.execute(
-            "SELECT value FROM meta WHERE key=?",
-            ("task_assignment_counter",),
-        ).fetchone()
+    provided_depth = data.get("depth")
+    parent_depth_value = None
+    if provided_depth is not None:
         try:
-            current_count = int(counter_row["value"]) if counter_row else 0
+            parent_depth_value = int(provided_depth)
         except (TypeError, ValueError):
-            current_count = 0
-        loop_count = current_count
-        while True:
-            next_count = loop_count + 1
-            refresh_due = next_count % 10 == 0
-            discovery_due = next_count % 100 == 0
-            checkpoint_due = next_count % 10000 == 0
-            should_truncate_wal = False
-
-            candidate_payload = None
-
-            if discovery_due:
-                candidate_payload = assign_discovery()
-                if candidate_payload is None and restart_discovery_cycle():
-                    should_truncate_wal = True
-                    candidate_payload = assign_discovery()
-
-            if candidate_payload is None and refresh_due:
-                assigned_row = retryable_execute(
-                    cur,
-                    """
-                    WITH candidate AS (
-                        SELECT steamAccountId
-                        FROM players
-                        WHERE hero_done=1
-                          AND assigned_to IS NULL
-                        ORDER BY COALESCE(hero_refreshed_at, '1970-01-01') ASC,
-                                 steamAccountId ASC
-                        LIMIT 1
-                    )
-                    UPDATE players
-                    SET hero_done=0,
-                        assigned_to='hero',
-                        assigned_at=CURRENT_TIMESTAMP
-                    WHERE steamAccountId IN (SELECT steamAccountId FROM candidate)
-                      AND hero_done=1
-                      AND assigned_to IS NULL
-                    RETURNING steamAccountId
-                    """,
-                ).fetchone()
-                if assigned_row:
-                    candidate_payload = {
-                        "type": "fetch_hero_stats",
-                        "steamAccountId": int(assigned_row["steamAccountId"]),
-                    }
-
-            if candidate_payload is None:
-                assigned_row = retryable_execute(
-                    cur,
-                    """
-                    WITH candidate AS (
-                        SELECT steamAccountId
-                        FROM players
-                        WHERE hero_done=0
-                          AND assigned_to IS NULL
-                        ORDER BY COALESCE(depth, 0) ASC, steamAccountId ASC
-                        LIMIT 1
-                    )
-                    UPDATE players
-                    SET assigned_to='hero',
-                        assigned_at=CURRENT_TIMESTAMP
-                    WHERE steamAccountId IN (SELECT steamAccountId FROM candidate)
-                      AND hero_done=0
-                      AND assigned_to IS NULL
-                    RETURNING steamAccountId
-                    """,
-                ).fetchone()
-                if assigned_row:
-                    candidate_payload = {
-                        "type": "fetch_hero_stats",
-                        "steamAccountId": int(assigned_row["steamAccountId"]),
-                    }
-
-            if candidate_payload is None:
-                hero_pending = cur.execute(
-                    "SELECT 1 FROM players WHERE hero_done=0 LIMIT 1"
-                ).fetchone()
-                if not hero_pending and not discovery_due:
-                    candidate_payload = assign_discovery()
-
-            if candidate_payload is not None:
-                task_payload = candidate_payload
-                retryable_execute(
-                    cur,
-                    """
-                    INSERT INTO meta (key, value)
-                    VALUES (?, ?)
-                    ON CONFLICT(key) DO UPDATE SET value=excluded.value
-                    """,
-                    ("task_assignment_counter", str(next_count)),
-                )
-                if checkpoint_due or should_truncate_wal:
-                    should_checkpoint = True
-                break
-
-            if refresh_due or discovery_due:
-                break
-
-            loop_count = next_count
-
-    if should_checkpoint:
-        with db_connection(write=True) as checkpoint_conn:
-            retryable_execute(
-                checkpoint_conn,
-                "PRAGMA wal_checkpoint(TRUNCATE);",
-            )
-
-    return task_payload
-
-
-def is_local_request() -> bool:
-    local_hosts = {"127.0.0.1", "::1"}
-    remote_addr = (request.remote_addr or "").strip()
-    if remote_addr in local_hosts or remote_addr.startswith("127."):
-        return True
-    for addr in request.access_route or []:
-        addr = (addr or "").strip()
-        if addr in local_hosts or addr.startswith("127."):
-            return True
-    forwarded_for = request.headers.get("X-Forwarded-For", "")
-    for addr in forwarded_for.split(","):
-        addr = addr.strip()
-        if addr and (addr in local_hosts or addr.startswith("127.")):
-            return True
-    return False
+            parent_depth_value = None
+    if parent_depth_value is None:
+        if assignment_row and assignment_row["depth"] is not None:
+            try:
+                parent_depth_value = int(assignment_row["depth"])
+            except (TypeError, ValueError):
+                parent_depth_value = 0
+        else:
+            parent_depth_value = 0
+    return parent_depth_value + 1
 
 
 def create_app() -> Flask:
@@ -502,49 +106,11 @@ def create_app() -> Flask:
         except (KeyError, TypeError, ValueError):
             return jsonify({"status": "error", "message": "steamAccountId is required"}), 400
         task_type = data.get("type")
-        with db_connection(write=True) as conn:
-            cur = conn.cursor()
-            if task_type == "fetch_hero_stats":
-                has_existing_stats = cur.execute(
-                    "SELECT 1 FROM hero_stats WHERE steamAccountId=? LIMIT 1",
-                    (steam_account_id,),
-                ).fetchone()
-                hero_done_value = 1 if has_existing_stats else 0
-                retryable_execute(
-                    cur,
-                    """
-                    UPDATE players
-                    SET hero_done=?,
-                        hero_refreshed_at=CASE WHEN ? THEN hero_refreshed_at ELSE NULL END,
-                        assigned_to=NULL,
-                        assigned_at=NULL
-                    WHERE steamAccountId=?
-                    """,
-                    (hero_done_value, hero_done_value, steam_account_id),
-                )
-            elif task_type == "discover_matches":
-                retryable_execute(
-                    cur,
-                    """
-                    UPDATE players
-                    SET discover_done=0,
-                        assigned_to=NULL,
-                        assigned_at=NULL
-                    WHERE steamAccountId=?
-                    """,
-                    (steam_account_id,),
-                )
-            else:
-                retryable_execute(
-                    cur,
-                    """
-                    UPDATE players
-                    SET assigned_to=NULL,
-                        assigned_at=NULL
-                    WHERE steamAccountId=?
-                    """,
-                    (steam_account_id,),
-                )
+        if not reset_player_task(steam_account_id, task_type):
+            return (
+                jsonify({"status": "error", "message": "Player not found"}),
+                404,
+            )
         return jsonify({"status": "ok"})
 
     @app.post("/submit")
@@ -557,27 +123,11 @@ def create_app() -> Flask:
                 steam_account_id = int(data["steamAccountId"])
             except (KeyError, TypeError, ValueError):
                 return jsonify({"status": "error", "message": "steamAccountId is required"}), 400
-            heroes_payload = data.get("heroes", [])
-            hero_stats_rows = []
-            best_rows = []
-            for hero in heroes_payload:
-                try:
-                    hero_id = int(hero["heroId"])
-                    matches_value = hero.get("matches", hero.get("games"))
-                    if matches_value is None:
-                        continue
-                    matches = int(matches_value)
-                    wins = int(hero.get("wins", 0))
-                except (KeyError, TypeError, ValueError):
-                    continue
-                hero_stats_rows.append((steam_account_id, hero_id, matches, wins))
-                hero_name = HEROES.get(hero_id)
-                if hero_name:
-                    best_rows.append(
-                        (hero_id, hero_name, steam_account_id, matches, wins)
-                    )
+            hero_stats_rows, best_rows = _extract_hero_rows(
+                steam_account_id,
+                data.get("heroes", []),
+            )
             assigned_at_value = None
-            update_count = 0
             with db_connection(write=True) as conn:
                 cur = conn.cursor()
                 assignment_row = retryable_execute(
@@ -598,19 +148,12 @@ def create_app() -> Flask:
                     """,
                     (steam_account_id,),
                 )
-                update_count = update_cursor.rowcount
-            if update_count == 0:
+            if update_cursor.rowcount == 0:
                 return (
-                    jsonify(
-                        {
-                            "status": "error",
-                            "message": "Player not found",
-                        }
-                    ),
+                    jsonify({"status": "error", "message": "Player not found"}),
                     404,
                 )
-            BACKGROUND_EXECUTOR.submit(
-                _process_hero_submission,
+            submit_hero_submission(
                 steam_account_id,
                 hero_stats_rows,
                 best_rows,
@@ -626,20 +169,8 @@ def create_app() -> Flask:
                 steam_account_id = int(data["steamAccountId"])
             except (KeyError, TypeError, ValueError):
                 return jsonify({"status": "error", "message": "steamAccountId is required"}), 400
-            discovered_ids: list[int] = []
-            seen_ids: set[int] = set()
-            for value in data.get("discovered", []):
-                try:
-                    candidate_id = int(value)
-                except (TypeError, ValueError):
-                    continue
-                if candidate_id in seen_ids:
-                    continue
-                seen_ids.add(candidate_id)
-                discovered_ids.append(candidate_id)
+            discovered_ids = _extract_discovered_ids(data.get("discovered", []))
             assigned_at_value = None
-            next_depth_value = None
-            update_count = 0
             with db_connection(write=True) as conn:
                 cur = conn.cursor()
                 assignment_row = retryable_execute(
@@ -649,29 +180,7 @@ def create_app() -> Flask:
                 ).fetchone()
                 if assignment_row is not None:
                     assigned_at_value = assignment_row["assigned_at"]
-                provided_next_depth = data.get("nextDepth")
-                if provided_next_depth is not None:
-                    try:
-                        next_depth_value = int(provided_next_depth)
-                    except (TypeError, ValueError):
-                        next_depth_value = None
-                if next_depth_value is None:
-                    provided_depth = data.get("depth")
-                    parent_depth_value = None
-                    if provided_depth is not None:
-                        try:
-                            parent_depth_value = int(provided_depth)
-                        except (TypeError, ValueError):
-                            parent_depth_value = None
-                    if parent_depth_value is None:
-                        if assignment_row and assignment_row["depth"] is not None:
-                            try:
-                                parent_depth_value = int(assignment_row["depth"])
-                            except (TypeError, ValueError):
-                                parent_depth_value = 0
-                        else:
-                            parent_depth_value = 0
-                    next_depth_value = parent_depth_value + 1
+                next_depth_value = _resolve_next_depth(data, assignment_row)
                 update_cursor = retryable_execute(
                     cur,
                     """
@@ -683,21 +192,12 @@ def create_app() -> Flask:
                     """,
                     (steam_account_id,),
                 )
-                update_count = update_cursor.rowcount
-            if update_count == 0:
+            if update_cursor.rowcount == 0:
                 return (
-                    jsonify(
-                        {
-                            "status": "error",
-                            "message": "Player not found",
-                        }
-                    ),
+                    jsonify({"status": "error", "message": "Player not found"}),
                     404,
                 )
-            if next_depth_value is None:
-                next_depth_value = 1
-            BACKGROUND_EXECUTOR.submit(
-                _process_discover_submission,
+            submit_discover_submission(
                 steam_account_id,
                 discovered_ids,
                 next_depth_value,
@@ -712,25 +212,7 @@ def create_app() -> Flask:
 
     @app.get("/progress")
     def progress():
-        with db_connection() as conn:
-            total = conn.execute("SELECT COUNT(*) AS c FROM players").fetchone()["c"]
-            hero_done = (
-                conn.execute(
-                    "SELECT COUNT(*) AS c FROM players WHERE hero_done=1"
-                ).fetchone()["c"]
-            )
-            discover_done = (
-                conn.execute(
-                    "SELECT COUNT(*) AS c FROM players WHERE discover_done=1"
-                ).fetchone()["c"]
-            )
-        return jsonify(
-            {
-                "players_total": total,
-                "hero_done": hero_done,
-                "discover_done": discover_done,
-            }
-        )
+        return jsonify(fetch_progress())
 
     @app.get("/seed")
     def seed():
@@ -743,50 +225,15 @@ def create_app() -> Flask:
             return Response("Use /seed?start=1&end=100", status=400)
         if end < start:
             return Response("End must be >= start", status=400)
-        with db_connection(write=True) as conn:
-            cur = conn.cursor()
-            for pid in range(start, end + 1):
-                retryable_execute(
-                    cur,
-                    """
-                    INSERT OR IGNORE INTO players (
-                        steamAccountId,
-                        depth,
-                        hero_done,
-                        discover_done
-                    )
-                    VALUES (?,?,0,0)
-                    """,
-                    (pid, 0),
-                )
+        seed_players(start, end)
         return jsonify({"seeded": [start, end]})
 
     @app.get("/leaderboards/<hero_slug>")
     def hero_leaderboard(hero_slug: str):
-        slug = hero_slug.strip().replace(" ", "_").lower()
-        hero_entry = HERO_SLUGS.get(slug)
-        if not hero_entry:
+        hero_payload = fetch_hero_leaderboard(hero_slug)
+        if hero_payload is None:
             abort(404)
-        hero_id, hero_name = hero_entry
-        with db_connection() as conn:
-            rows = conn.execute(
-                """
-                SELECT steamAccountId, matches, wins
-                FROM hero_stats
-                WHERE heroId=?
-                ORDER BY matches DESC, wins DESC, steamAccountId ASC
-                LIMIT 100
-                """,
-                (hero_id,),
-            ).fetchall()
-        players = [
-            {
-                "steamAccountId": row["steamAccountId"],
-                "matches": row["matches"],
-                "wins": row["wins"],
-            }
-            for row in rows
-        ]
+        hero_name, slug, players = hero_payload
         return render_template(
             "leaderboard.html",
             hero_name=hero_name,
@@ -796,14 +243,6 @@ def create_app() -> Flask:
 
     @app.get("/best")
     def best():
-        with db_connection() as conn:
-            rows = conn.execute("SELECT * FROM best ORDER BY matches DESC").fetchall()
-        payload = []
-        for row in rows:
-            row_dict = dict(row)
-            name = row_dict.get("hero_name")
-            row_dict["hero_slug"] = hero_slug(name) if isinstance(name, str) else None
-            payload.append(row_dict)
-        return jsonify(payload)
+        return jsonify(fetch_best_payload())
 
     return app
